+++ core/alsatian-core.ts
@@ -5,7 +5,15 @@ import {
 } from "./";
 
 import {
-   Matcher
+    ContainerMatcher,
+    EmptyMatcher,
+    FunctionMatcher,
+    FunctionSpyMatcher,
+    Matcher,
+    MixedMatcher,
+    NumberMatcher,
+    PropertyMatcher,
+    StringMatcher
 } from "./matchers";
 
 import {
@@ -61,32 +69,40 @@ export {
    AsyncTeardown,
    AsyncTeardownFixture,
    AsyncTest,
+   ContainerMatcher,
+   EmptyMatcher,
    Expect,
    FocusTest,
    FocusTests,
+   FunctionMatcher,
+   FunctionSpy,
+   FunctionSpyMatcher,
    IgnoreTest,
    IgnoreTests,
    Matcher,
+   MatchError,
+   METADATA_KEYS,
+   MixedMatcher,
+   NumberMatcher,
+   PropertyMatcher,
    Setup,
    SetupFixture,
-   FunctionSpy,
    SpyOn,
    SpyOnProperty,
+   StringMatcher,
    Teardown,
    TeardownFixture,
    Test,
    TestCase,
-   TestRunner,
-   TestSet,
-   Timeout,
-   TestSetResults,
-   TestOutcome,
-   METADATA_KEYS,
-   TestFixture,
-   TestOutputStream,
-   MatchError,
-   TestFixtureResults,
-   TestResults,
    TestCaseResult,
-   TestTimeoutError
+   TestFixture,
+   TestFixtureResults,
+   TestOutcome,
+   TestOutputStream,
+   TestResults,
+   TestRunner,
+   TestSetResults,
+   TestSet,
+   TestTimeoutError,
+   Timeout
 };
