+++ core/matchers/index.ts
@@ -4,6 +4,7 @@ import { FunctionMatcher } from "./function-matcher";
 import { FunctionSpyCallCountMatcher } from "./function-spy-call-count-matcher";
 import { FunctionSpyMatcher } from "./function-spy-matcher";
 import { Matcher } from "./matcher";
+import { MixedMatcher } from "./mixed-matcher";
 import { NumberMatcher } from "./number-matcher";
 import { PropertyMatcher } from "./property-matcher";
 import { SpyCallCountType } from "./spy-call-count-type";
@@ -16,6 +17,7 @@ export {
    FunctionSpyCallCountMatcher,
    FunctionSpyMatcher,
    Matcher,
+   MixedMatcher,
    NumberMatcher,
    PropertyMatcher,
    SpyCallCountType,
